chore: refactor sidebar.js to use declarative format
Docusauraus provides a short hand format that is more
convenient but harder to read as a relies on more familiarity
with the conventions of the documentation tool.

We should optimise for readability and use their more verbose
format. This will make it more straight forward for new
contributors to edit the documentation site.