feat: adds logger lib
Before the addition of this tool all logs
within the primary client side application
are being done via console.log. This is
okay for local development but makes it
harder to investigate issues affecting other
people as we are not currently piping this data
to anywhere.

Pino is a performant log tool that will
allow us to configure additional destinatons for
these logs. We could easily aggregate logs data to
a 3rd party tool based on the configuration of a specific
deployment.

For the time being this logger will be hardcoded to output
logs to browser console. This is to avoid any disruption to
expected dev behaviour.