feat: load sentry DSN from env
The implementation should support
Sentry credentials being supplied via the
environment.

This allows us to use seperate Sentry projects
for individual installs and eventually remove
the credentials from our codebase entirely
instead moving this key to be supplied via
CI environment variables.

In addition to this change we are adding
support for Sentry's concept of environments.
We are using the `siteVariant` to acheive this.

This allows us to filter the errors and understand
how our production environment is performing compared
to our various staging and development environment.