feat: load configuration from outside of core app bundle
A full build of the application is needed for each target installation
because the credentials needed by Firebase, Algolia, Sentry etc are
loaded out of the environmental variables.

Builds require multiple minutes to run and this work is duplicative.
Recently we had to upgrade the CircleCI resources used to sidestep
performance issues around the build step.

Duplicate builds mean that the code which we deploy to our production
environment is not the same as the code under test.