test: introduce coverage for research routes
The order of items is important for the router,
up until now there were no tests which validated
that the correct components were being loaded
based on the paths available.

Alongside introducting this tests some of
the components within the rendering tree
have been adjusted to make them more tolerant
of partially available store data. This is a
lazy alternative which allows us to instantiate
a minimal amount of items and avoids us creating
tests which are too brittle.