fix: service worker cache headers
Over the last month we've shipped multiple
releases to production. At a far higher
cadence then before.

A side effect of this has been observing
multiple "odd" behaviour on the production
deployments which can not be recreated locally.

The primary candidate for causing these issues
is the service worker. The service worker is a
built in component of create react app (CRA).

Interestingly in the CRA documentation when discussing
servic workers they mention:
> they (service workers) can make debugging deployments more challenging.

So as a result I have been hunting out for more information around
other folks encountering this behaviour.

An interesting thread here around the combination of
service workers AND firebase:
https://github.com/facebook/create-react-app/issues/2440#issuecomment-305592301