test: increase timeout
This test has been flaking a little bit in Circle
recently. Looking at the resource usage the small
Docker resource is running at 100% CPU. We have
also been introducing _more_ unit tests into the
platform.

This is an experiment to see if this stabilises
the tests, but it may be we need to consider either:

* Increasing the resource for running tests
* Dig into jest configuration inherited from create
react app to identify optimisations.