feat: removes legacy mail function
A legacy function exists that served mostly as a proof of concept to
automate sending emails from the platform based on db updates.

As the feature is outdated, not currently used, and requires additional
setup from firebase console to enable integration and mail provider,
assume easiest for now to just remove.

BREAKING CHANGE: The function will need to be manually deleted across all current instances