chore(ci): remove firefox e2e tests
The Cypress tests are now consistently failing
when running against the Firefox browser. However
driving the browser through the same steps it
is not possible to encounter the same issue.

This _appears_ to be a false positive, however
it is now blocking the production pipeline.

This change aims to remove this blockage so we
can debug this issue around Cypress + Firefox
stability separately.