fix: increase caching for static JS resources
After deploying a new iteration of the site a user
can hit an error state due to the service worker
attempting to load a JS file which no longer exists
on the Firebase hosting. Each deployment removes
all previous files from the hosting.

The default caching is 1 hour. However as the JS
bundles include a md5 string they are unique so can
be cached indefinitely.