feat: refactor to remove aggregations dependency from comments
The VerifiedUserBadge component introduces a dependency
on the aggregations store. This is a useful helper function however
we should be able to pre-calculate this when fetching the comment data
for a ResearchUpadte or HowTo article.

As a result of this change it is now possible to extract the Comment and
CommentsList components out to the components package. This makes
documentation and testing edge cases easier.