fix: remove service-worker caching for index
The service worker should not cache the index.html page as
this is the entry point to our service which contains
the manifest for which JS files to load.

After each deployment, the old JS files are removed from the
server. As a result the index.html loaded from cache by the service
worker can include references to JS files which no longer exist.