feat: extract create comment component
Refactors existing `CommentTextArea` component into shared
`CreateComment` component which groups the user avatar and
submission button.

This change also improves the mobile rendering which removes
the need for #1650.

Whilst I think there is still further improvement to be made
on this component for a mobile user that can be done in a
future iteration with oversight from the Design team.

Closes #1737

wip

stash