chore(ci): generate releases on master(default) branch
Originally the releases were generated _after_ the code
had been deployed to a production environment. However
as the number of deployed instances has increased past one,
the question is if code is partially deployed has been released?

Not wishing to get lost down that rabbit hole we opted to declare
code released _before_ it is then deployed to the instances.