feat: introduce @mention to howto and research items
* When @mention a user with their username, a UserReference
will be stored within the DB that contains the authId AND username.
* When loading the data out of the DB the UserReference string is
converted to a plain `@mention` string.
* Switches in a new plugin for automatically adding links to text. This
new plugin adds support for @mention links. As a result `@mention` will
now link through to `/u/mention`.

Notifications are NOT handled as part of this change. This means that
no users will be notified if someone else `@mention` them, that work
will be addressed in another PR.