feat: supports contributors on research item #2006
Introduces an array item to Research articles which contains a list
of usernames. This is a deliberately naive implementation as Research
articles are currently only available on the Project Kamp instance.

Future development here may involve aligning with the UserReference
pattern used to persist @mention references. The UserReference
convention is more robust as it uses IDs rather than usernames to
identify users. The additional complexity is difficult to justify
in this instance as the target user story here only covers a few
Project Kamp contributors.