chore: refactor notification message out of components
The original implementation relied heavily on the component library
knowing about the structure of the notification object. The
interface has been simplified so that all it does it use type to
determine the icon and otherwise wrap the passed content.

A local theme context has been introduced to style any
nested anchor elements to match the original design.