chore(lint): make rules stricter
This rule is configured somewhere as a warning, which
will fail the build step.

> Creating an optimized production build...
>
> Treating warnings as errors because process.env.CI = true.
> Most CI servers set it automatically.

This configuration is likely within the create react app
project. Failing on unused variables is a useful
configuration to adopt as it indicates incomplete changes
and is not something a human reviewer should need to
think about when looking at PRs.