feat: expose logging method via routing
At the moment the serverless functions are
only available via a hard to discover URL this
means they can not be found at a root relative path.

This change exposes them via a temporary URL for testing.
Opinions welcome on whether this should eventually be
exposed via a namespaced URI such as `/api/logging`