feat: load country information off of User
The research author country information was being stored
on the document rather than fetched from the User. The original
design aim here was to avoid reads from the User document as Firebase
charges for each document read.

However this optimisation resulted in a suboptimal user experience
as country information on Research articles failed to stay in sync
with the User document.

This change is intended to introduce _more_ reads to the system
which may increase the operational costs but reduce the application
complexity and improve the reader experience.