feat: add header (search, filter, sort) to questions
Updated SortFilterHeader to take the type question and render category filter, sort dropdown and
search box. Updated FilterSorterDecorator to include as ItemSortingOption:  LeastComments and
LatestComments. Updated question.store to be able to search, sort and filter. Added to
QuestionListing the SortFilterHeader. Created new file src/stores/utils.ts with a function that
checks if a store has author functionality (if not then SortFilterHeader does not render Filter by
author dropdown"

closes #2969, #2968, #2967